Better display times, prometheus install
