+++ server/go.mod
@@ -5,5 +5,6 @@ go 1.16
 require (
 	github.com/gorilla/mux v1.8.0
 	github.com/lib/pq v1.10.0 // indirect
+	github.com/prometheus/client_golang v1.10.0 // indirect
 	github.com/stretchr/testify v1.7.0 // indirect
 )
